Rename HP# to HP-BRCU
