+++ Cargo.toml
@@ -16,7 +16,7 @@ num = "0.4.1"
 hp_pp = { path = "./hp_pp" }
 nbr-rs = { path = "./nbr-rs" }
 cdrc-rs = { path = "./cdrc-rs" }
-hp-sharp = { path = "./hp-sharp" }
+hp-brcu = { path = "./hp-brcu" }
 vbr-rs = { path = "./vbr-rs" }
 circ = { path = "./circ" }
