+++ hp-brcu/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hp-sharp"
+name = "hp-brcu"
 version = "0.1.0"
 edition = "2021"
