+++ hp-brcu/sanitize-apple.sh
