+++ hp-brcu/src/deferred.rs
