+++ hp-brcu/src/hazard.rs
