+++ src/ds_impl/mod.rs
@@ -4,7 +4,7 @@ pub mod circ_hp;
 pub mod ebr;
 pub mod hp;
 pub mod hp_pp;
-pub mod hp_sharp;
+pub mod hp_brcu;
 pub mod nbr;
 pub mod nr;
 pub mod pebr;
